Upgraded trajectory generation

+++ sw/src/Timer5.c
@@ -2,7 +2,7 @@
 #include <Arduino.h>
 #include "Timer5.h"
 
-void setTimer5 (uint16_t interval); {
+void setTimer5 (uint32_t interval); {
 
 	/*
 				
@@ -24,7 +24,7 @@ void setTimer5 (uint16_t interval); {
 	TCCR5B = 0;
 	TCNT5  = 0;
 
-	OCR5A = (interval/64) - 1; // compare match register 16MHz clock, 1 prescaler, 400kHz frequency (the timer stats at zero hence the -1)
+	OCR5A = (uint16_t) (interval/64) - 1; // compare match register 16MHz clock, 1 prescaler, 400kHz frequency (the timer stats at zero hence the -1)
 	
 	// CTC mode
 	TCCR5B |= (1 << WGM12);
